move Firestore operation types to types folder
