+++ src/hooks/albums/useAlbums.ts
@@ -1,5 +1,5 @@
 import { getAlbumDocs } from "@/data/firestore/albums";
-import { QueryOptions } from "@/data/firestore/firestoreClientOperations";
+import { QueryOptions } from "@/data/firestore/types/queries";
 import { AlbumDoc } from "@/data/firestore/types/documents";
 import { Album } from "@/types/albums/albumTypes";
 import { useInfiniteQuery, useQueryClient } from "@tanstack/react-query";
